LayoutTests/media/video-restricted-invisible-autoplay-allowed-after-play-expected.txt

Test that "invisible autoplay not allowed restriction" does not pause media after an explicit call to play().


** setting video.src
EVENT(play)
Promise resolved OK
END OF TEST
